@@ -1,15 +1,36 @@
 cmake_minimum_required(VERSION 3.20)
 
-set(SlopeCraft_version 5.0.0)
+# set version -----------------------------------------------------------------
+set(SlopeCraft_version 5.3.0)
 
+# set basic project attributes ------------------------------------------------
 project(SlopeCraft VERSION ${SlopeCraft_version} LANGUAGES C CXX)
-set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD 23)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 set(CMAKE_VISIBILITY_INLINES_HIDDEN TRUE)
 
-set(SlopeCraft_GPU_API "OpenCL" CACHE STRING "API used to compute. Valid values : OpenCL, None. Metal may be supported.")
+if (${LINUX})
+    message(STATUS "setting runpath for linux")
+    set(CMAKE_INSTALL_RPATH "$ORIGIN/../lib")
 
-option(SlopeCraft_vectorize "Compile with vectorization" ON)
+    # set(CMAKE_BUILD_WITH_INSTALL_RPATH TRUE)
+    # set(CMAKE_INSTALL_RPATH_USE_LINK_PATH FALSE)
+endif ()
+
+enable_testing()
+
+# configure options -----------------------------------------------------------
+if (${APPLE})
+    set(SlopeCraft_GPU_API "None" CACHE STRING "API used to compute. Valid values : OpenCL, None. Metal may be supported.")
+    option(SlopeCraft_vectorize "Compile with vectorization" OFF)
+    message(STATUS "GPU boosting and vectorization have been disabled on mac by default")
+
+else ()
+    set(SlopeCraft_GPU_API "OpenCL" CACHE STRING "API used to compute. Valid values : OpenCL, None. Metal may be supported.")
+    option(SlopeCraft_vectorize "Compile with vectorization" ON)
+endif ()
+
 
 option(SlopeCraft_update_ts_files "Update language files before build. If it is set to ON, everytime CMake reconfigure the project, all .ts files will be updated." OFF)
 
