@@ -41,7 +41,7 @@
     </message>
     <message>
         <location filename="MapViewerWind.ui" line="150"/>
-        <location filename="MapViewerWind.cpp" line="256"/>
+        <location filename="MapViewerWind.cpp" line="253"/>
         <source>请选择地图文件</source>
         <translation>Please select files</translation>
     </message>
@@ -97,8 +97,8 @@
     </message>
     <message>
         <location filename="MapViewerWind.ui" line="408"/>
-        <location filename="MapViewerWind.cpp" line="623"/>
-        <location filename="MapViewerWind.cpp" line="637"/>
+        <location filename="MapViewerWind.cpp" line="619"/>
+        <location filename="MapViewerWind.cpp" line="632"/>
         <source>保存为图片</source>
         <translation>Save</translation>
     </message>
@@ -108,43 +108,43 @@
         <translation>Show grid</translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="197"/>
+        <location filename="MapViewerWind.cpp" line="196"/>
         <source>地图数：</source>
         <translation>Map count: </translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="371"/>
+        <location filename="MapViewerWind.cpp" line="368"/>
         <source>有</source>
         <translation>There is/are </translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="372"/>
+        <location filename="MapViewerWind.cpp" line="369"/>
         <source>个像素点出现未知的基色，它们使用了Mojang未定义的基色。这可能是因为软件版本较旧而游戏版本太新，或者地图文件损坏。
 其行、列坐标分别为：</source>
         <translation> pixel(s) which have unknown base color, they are using basecolors out of Mojang&apos;s defination. This may be caused by damaged map data file, or the software version is too old.</translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="382"/>
+        <location filename="MapViewerWind.cpp" line="379"/>
         <source>地图中有Mojang未定义基色</source>
         <translation>Undefined basecolor in map data files (undefined by Mojang)</translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="534"/>
+        <location filename="MapViewerWind.cpp" line="530"/>
         <source>选择地图数据文件</source>
         <translation>Select map data files</translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="573"/>
+        <location filename="MapViewerWind.cpp" line="569"/>
         <source>加载地图文件失败</source>
         <translation>Failed to load map data file(s)</translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="574"/>
+        <location filename="MapViewerWind.cpp" line="570"/>
         <source>出错的文件：</source>
         <translation>Map data file: </translation>
     </message>
     <message>
-        <location filename="MapViewerWind.cpp" line="574"/>
+        <location filename="MapViewerWind.cpp" line="570"/>
         <source>
 错误信息：
 </source>
