@@ -26,9 +26,9 @@ This file is part of SlopeCraft.
 #include <magic_enum.hpp>
 #include <map>
 
-const VCL_block *
-find_first_mark_block(const std::unordered_map<const VCL_block *, uint16_t>
-                          &blocks_allowed) noexcept {
+const VCL_block *find_first_mark_block(
+    const std::unordered_map<const VCL_block *, uint16_t>
+        &blocks_allowed) noexcept {
   const VCL_block *p = nullptr;
   for (const auto &pair : blocks_allowed) {
     if (pair.first->is_air()) {
@@ -46,7 +46,6 @@ find_first_mark_block(const std::unordered_map<const VCL_block *, uint16_t>
 }
 
 bool TokiVC::export_test_litematic_no_lock(const char *filename) noexcept {
-
   if (!TokiVC_internal::is_basic_color_set_ready ||
       !TokiVC_internal::is_allowed_color_set_ready) {
     VCL_report(
@@ -142,6 +141,6 @@ bool TokiVC::export_test_litematic_no_lock(const char *filename) noexcept {
     info.author_utf8 = "VisualCraftL";
     info.destricption_utf8 = "This litematic is generated by VisualCraft.";
 
-    return schem.export_litematic(filename, info);
+    return schem.export_litematic(filename, info, nullptr, nullptr);
   }
 }
