@@ -23,8 +23,10 @@ This file is part of SlopeCraft.
 #include "TokiVC.h"
 #include "VCL_internal.h"
 #include "VisualCraftL.h"
-#include <immintrin.h>
-#include <xmmintrin.h>
+
+#ifdef min
+#undef min
+#endif
 
 constexpr uint32_t reverse_color(uint32_t ARGB_src) noexcept {
   return ARGB32(255 - getR(ARGB_src), 255 - getG(ARGB_src),
@@ -235,9 +237,9 @@ bool TokiVC::export_flag_diagram(const char *png_filename,
             "a subproject of SlopeCraft, developed by TokiNoBug."),
         std::make_pair<std::string, std::string>(
             "Comment",
-            "VisualCraftL is a free software published "
+            "SlopeCraft is a free software published "
             "under GPLv3 license. You can find "
-            "its repository at https://github.com/ToKiNoBug/SlopeCraft")};
+            "its repository at https://github.com/SlopeCraft/SlopeCraft")};
 
     std::array<png_text, txt.size()> png_txts;
 
