@@ -177,6 +177,17 @@ enum class VCL_upper_direction_t : uint8_t {
 };
 */
 
+struct VCL_string_deliver {
+  char *data{nullptr};
+  size_t size{0};
+  size_t capacity{0};
+};
+
+struct VCL_read_only_buffer {
+  const void *data{nullptr};
+  size_t size{0};
+};
+
 class VCL_Kernel;
 class VCL_resource_pack;
 class VCL_block_state_list;
@@ -187,7 +198,7 @@ class VCL_GPU_Platform;
 class VCL_GPU_Device;
 
 class VCL_Kernel {
-public:
+ public:
   VCL_Kernel() = default;
   virtual ~VCL_Kernel() = default;
 
@@ -217,9 +228,9 @@ public:
   virtual int64_t cols() const noexcept = 0;
   // virtual int64_t layers() const noexcept = 0;
 
-  virtual const uint32_t *
-  raw_image(int64_t *const rows, int64_t *const cols,
-            bool *const is_row_major) const noexcept = 0;
+  virtual const uint32_t *raw_image(
+      int64_t *const rows, int64_t *const cols,
+      bool *const is_row_major) const noexcept = 0;
 
   ////////////////////////////////////////////////////////////////////////
   virtual bool convert(::SCL_convertAlgo algo,
@@ -233,8 +244,8 @@ public:
     // [row_start,row_end) * [o,col_count) will be written
     int64_t row_start;
     int64_t row_end;
-    int32_t split_line_row_margin; // 0 or negative number means no split lines
-    int32_t split_line_col_margin; // 0 or negative number means no split lines
+    int32_t split_line_row_margin;  // 0 or negative number means no split lines
+    int32_t split_line_col_margin;  // 0 or negative number means no split lines
     int png_compress_level{9};
     int png_compress_memory_level{8};
   };
@@ -264,6 +275,14 @@ public:
       const int (&weOffset)[3] = {0, 0, 0}, const char *utf8_Name = "",
       const char *const *const utf8_requiredMods = nullptr,
       const int requiredModsCount = 0) const noexcept = 0;
+
+  struct gpu_options {
+    const uint64_t lib_version{SC_VERSION_U64};
+    VCL_string_deliver *error_message{nullptr};
+  };
+  virtual bool set_gpu_resource(const VCL_GPU_Platform *,
+                                const VCL_GPU_Device *,
+                                const gpu_options &option) noexcept = 0;
   /*
 /// export map into Structure files (*.NBT)
 virtual void exportAsWESchem(
@@ -281,18 +300,21 @@ extern "C" {
 VCL_EXPORT_FUN void VCL_destroy_kernel(VCL_Kernel *const ptr);
 
 // create and destroy resource pack
+[[nodiscard]] VCL_EXPORT_FUN VCL_resource_pack *VCL_create_resource_pack(
+    const int zip_file_count, const char *const *const zip_file_names);
 [[nodiscard]] VCL_EXPORT_FUN VCL_resource_pack *
-VCL_create_resource_pack(const int zip_file_count,
-                         const char *const *const zip_file_names);
-VCL_EXPORT_FUN void VCL_destroy_resource_pack(VCL_resource_pack *const ptr);
+VCL_create_resource_pack_from_buffers(const size_t zip_count,
+                                      const VCL_read_only_buffer *file_contents,
+                                      const char *const *const zip_file_names);
+VCL_EXPORT_FUN
+void VCL_destroy_resource_pack(VCL_resource_pack *const ptr);
 
 // create and destroy block state list
-[[nodiscard]] VCL_EXPORT_FUN VCL_block_state_list *
-VCL_create_block_state_list(const int file_count,
-                            const char *const *const json_file_names);
+[[nodiscard]] VCL_EXPORT_FUN VCL_block_state_list *VCL_create_block_state_list(
+    const int file_count, const char *const *const json_file_names);
 
-VCL_EXPORT_FUN void
-VCL_destroy_block_state_list(VCL_block_state_list *const ptr);
+VCL_EXPORT_FUN void VCL_destroy_block_state_list(
+    VCL_block_state_list *const ptr);
 
 struct VCL_set_resource_option {
   const uint64_t lib_version{SC_VERSION_U64};
@@ -303,21 +325,18 @@ struct VCL_set_resource_option {
   bool is_render_quality_fast;
 };
 
-VCL_EXPORT_FUN double
-VCL_estimate_color_num(size_t num_layers, size_t num_foreground,
-                       size_t num_background,
-                       size_t num_nontransparent_non_background);
+VCL_EXPORT_FUN double VCL_estimate_color_num(
+    size_t num_layers, size_t num_foreground, size_t num_background,
+    size_t num_nontransparent_non_background);
 
 // set resource for kernel
-VCL_EXPORT_FUN bool
-VCL_set_resource_copy(const VCL_resource_pack *const rp,
-                      const VCL_block_state_list *const bsl,
-                      const VCL_set_resource_option &option);
+VCL_EXPORT_FUN bool VCL_set_resource_copy(
+    const VCL_resource_pack *const rp, const VCL_block_state_list *const bsl,
+    const VCL_set_resource_option &option);
 
-VCL_EXPORT_FUN bool
-VCL_set_resource_move(VCL_resource_pack **rp_ptr,
-                      VCL_block_state_list **bsl_ptr,
-                      const VCL_set_resource_option &option);
+VCL_EXPORT_FUN bool VCL_set_resource_move(
+    VCL_resource_pack **rp_ptr, VCL_block_state_list **bsl_ptr,
+    const VCL_set_resource_option &option);
 
 VCL_EXPORT_FUN void VCL_discard_resource();
 
@@ -332,15 +351,13 @@ VCL_EXPORT_FUN size_t VCL_num_basic_colors();
 /**
   \returns the number of blocks of this color.
 */
-VCL_EXPORT_FUN int
-VCL_get_basic_color_composition(size_t color_idx,
-                                const VCL_block **const blocks_dest = nullptr,
-                                uint32_t *const color_dest = nullptr);
+VCL_EXPORT_FUN int VCL_get_basic_color_composition(
+    size_t color_idx, const VCL_block **const blocks_dest = nullptr,
+    uint32_t *const color_dest = nullptr);
 
 // set allowed blocks for kernel
-VCL_EXPORT_FUN bool
-VCL_set_allowed_blocks(const VCL_block *const *const blocks_allowed,
-                       size_t num_block_allowed);
+VCL_EXPORT_FUN bool VCL_set_allowed_blocks(
+    const VCL_block *const *const blocks_allowed, size_t num_block_allowed);
 VCL_EXPORT_FUN void VCL_discard_allowed_blocks();
 
 VCL_EXPORT_FUN bool VCL_is_allowed_colorset_ok();
@@ -425,12 +442,11 @@ VCL_EXPORT_FUN bool VCL_compare_block(const VCL_block *b1, const VCL_block *b2);
 VCL_EXPORT_FUN VCL_block_class_t VCL_string_to_block_class(const char *str,
                                                            bool *ok = nullptr);
 
-[[nodiscard]] VCL_EXPORT_FUN VCL_model *
-VCL_get_block_model(const VCL_block *block,
-                    const VCL_resource_pack *resource_pack);
+[[nodiscard]] VCL_EXPORT_FUN VCL_model *VCL_get_block_model(
+    const VCL_block *block, const VCL_resource_pack *resource_pack);
 
-[[nodiscard]] VCL_EXPORT_FUN VCL_model *
-VCL_get_block_model_by_name(const VCL_resource_pack *, const char *name);
+[[nodiscard]] VCL_EXPORT_FUN VCL_model *VCL_get_block_model_by_name(
+    const VCL_resource_pack *, const char *name);
 
 VCL_EXPORT_FUN void VCL_destroy_block_model(VCL_model *);
 
@@ -463,8 +479,8 @@ i =
   if other values, the function will return INT_MIN.
 */
 VCL_EXPORT_FUN int VCL_version_component(int i);
-VCL_EXPORT_FUN bool
-VCL_is_version_ok(uint64_t version_at_caller_s_build_time = SC_VERSION_U64);
+VCL_EXPORT_FUN bool VCL_is_version_ok(
+    uint64_t version_at_caller_s_build_time = SC_VERSION_U64);
 
 VCL_EXPORT_FUN bool VCL_have_gpu_api();
 VCL_EXPORT_FUN const char *VCL_get_GPU_api_name();
@@ -476,9 +492,8 @@ VCL_EXPORT_FUN void VCL_release_platform(VCL_GPU_Platform *);
 VCL_EXPORT_FUN const char *VCL_get_platform_name(const VCL_GPU_Platform *);
 
 VCL_EXPORT_FUN size_t VCL_get_device_num(const VCL_GPU_Platform *);
-[[nodiscard]] VCL_EXPORT_FUN VCL_GPU_Device *
-VCL_get_device(const VCL_GPU_Platform *, size_t device_idx,
-               int *errorcode = nullptr);
+[[nodiscard]] VCL_EXPORT_FUN VCL_GPU_Device *VCL_get_device(
+    const VCL_GPU_Platform *, size_t device_idx, int *errorcode = nullptr);
 VCL_EXPORT_FUN void VCL_release_device(VCL_GPU_Device *);
 VCL_EXPORT_FUN const char *VCL_get_device_name(const VCL_GPU_Device *);
 
@@ -493,6 +508,28 @@ VCL_EXPORT_FUN const char *VCL_biome_name(VCL_biome_t, uint8_t is_ZH);
 VCL_EXPORT_FUN uint32_t VCL_locate_colormap(const VCL_resource_pack *,
                                             bool is_grass, VCL_biome_info info,
                                             int *row, int *col);
+
+class VCL_preset;
+
+VCL_EXPORT VCL_preset *VCL_create_preset();
+VCL_EXPORT VCL_preset *VCL_load_preset(const char *filename,
+                                       VCL_string_deliver *error);
+VCL_EXPORT bool VCL_save_preset(const VCL_preset *, const char *filename,
+                                VCL_string_deliver *error);
+VCL_EXPORT void VCL_destroy_preset(VCL_preset *);
+VCL_EXPORT bool VCL_preset_contains_id(const VCL_preset *, const char *id);
+VCL_EXPORT void VCL_preset_emplace_id(VCL_preset *, const char *id);
+VCL_EXPORT bool VCL_preset_contains_class(const VCL_preset *,
+                                          VCL_block_class_t);
+VCL_EXPORT void VCL_preset_emplace_class(VCL_preset *, VCL_block_class_t);
+VCL_EXPORT size_t VCL_preset_num_ids(const VCL_preset *p);
+VCL_EXPORT size_t VCL_preset_get_ids(const VCL_preset *p, const char **id_dest,
+                                     size_t capacity);
+VCL_EXPORT size_t VCL_preset_num_classes(const VCL_preset *p);
+VCL_EXPORT size_t VCL_preset_get_classes(const VCL_preset *p,
+                                         VCL_block_class_t *class_dest,
+                                         size_t capacity);
+VCL_EXPORT void VCL_preset_clear(VCL_preset *);
 }
 
-#endif // SLOPECRAFT_VISUALCRAFT_VISUALCRAFT_H
+#endif  // SLOPECRAFT_VISUALCRAFT_VISUALCRAFT_H
