@@ -23,10 +23,15 @@ This file is part of SlopeCraft.
 #ifndef SLOPECRAFT_VISUALCRAFT_VISUALCRAFTL_GLOBAL_H
 #define SLOPECRAFT_VISUALCRAFT_VISUALCRAFTL_GLOBAL_H
 
-// #ifdef VISUALCRAFTL_BUILD
+#ifdef Q_DECL_EXPORT
+#undef Q_DECL_EXPORT
+#endif
+#ifdef Q_DECL_IMPORT
+#undef Q_DECL_IMPORT
+#endif
 
-#if defined(_MSC_VER) || defined(WIN64) || defined(_WIN64) ||                  \
-    defined(__WIN64__) || defined(WIN32) || defined(_WIN32) ||                 \
+#if defined(_MSC_VER) || defined(WIN64) || defined(_WIN64) ||  \
+    defined(__WIN64__) || defined(WIN32) || defined(_WIN32) || \
     defined(__WIN32__) || defined(__NT__)
 #define Q_DECL_EXPORT __declspec(dllexport)
 #define Q_DECL_IMPORT __declspec(dllimport)
@@ -42,7 +47,7 @@ This file is part of SlopeCraft.
 #endif
 
 #ifdef _WIN32
-#define VCL_EXPORT_FUN __stdcall VCL_EXPORT
+#define VCL_EXPORT_FUN VCL_EXPORT
 #else
 #define VCL_EXPORT_FUN VCL_EXPORT
 #endif
@@ -53,4 +58,4 @@ This file is part of SlopeCraft.
 #include "SC_version_buildtime.h"
 #endif
 
-#endif // SLOPECRAFT_VISUALCRAFT_VISUALCRAFTL_GLOBAL_H
+#endif  // SLOPECRAFT_VISUALCRAFT_VISUALCRAFTL_GLOBAL_H
