@@ -30,7 +30,6 @@ This file is part of SlopeCraft.
 using std::cout, std::endl;
 
 int main(int argc, char **argv) {
-
   CLI::App app;
 
   std::vector<std::string> input_files;
@@ -53,7 +52,7 @@ int main(int argc, char **argv) {
 int __version;
 app.add_option("--version", __version, "MC version.")
     ->default_val(19)
-    ->check(CLI::Range(12, 19, "Avaliable versions."));
+    ->check(CLI::Range(12, 20, "Avaliable versions."));
 
     */
     /*
