@@ -33,6 +33,7 @@ This file is part of SlopeCraft.
 #include <thread>
 #include <type_traits>
 #include <unordered_map>
+#include <cereal/cereal.hpp>
 
 #include "../SC_GlobalEnums.h"
 #include "ColorManip.h"
@@ -41,30 +42,31 @@ This file is part of SlopeCraft.
 
 #ifdef RGB
 #undef RGB
-#endif // #ifdef RGB
+#endif  // #ifdef RGB
 
 namespace libImageCvt {
 
 using ::Eigen::Dynamic;
 
-static const Eigen::Array<float, 2, 3>
-    dithermap_LR({{0.0 / 16.0, 0.0 / 16.0, 7.0 / 16.0},
-                  {3.0 / 16.0, 5.0 / 16.0, 1.0 / 16.0}});
+static const Eigen::Array<float, 2, 3> dithermap_LR(
+    {{0.0 / 16.0, 0.0 / 16.0, 7.0 / 16.0},
+     {3.0 / 16.0, 5.0 / 16.0, 1.0 / 16.0}});
 
-static const Eigen::Array<float, 2, 3>
-    dithermap_RL({{7.0 / 16.0, 0.0 / 16.0, 0.0 / 16.0},
-                  {1.0 / 16.0, 5.0 / 16.0, 3.0 / 16.0}});
+static const Eigen::Array<float, 2, 3> dithermap_RL(
+    {{7.0 / 16.0, 0.0 / 16.0, 0.0 / 16.0},
+     {1.0 / 16.0, 5.0 / 16.0, 3.0 / 16.0}});
 
-template <bool is_not_optical> struct GPU_wrapper_wrapper {
+template <bool is_not_optical>
+struct GPU_wrapper_wrapper {
   constexpr bool have_gpu_resource() const noexcept { return false; }
 };
 
-template <> struct GPU_wrapper_wrapper<false> {
-
-protected:
+template <>
+struct GPU_wrapper_wrapper<false> {
+ protected:
   gpu_wrapper::gpu_interface *gpu{nullptr};
 
-public:
+ public:
   bool have_gpu_resource() const noexcept { return this->gpu != nullptr; }
 
   void set_gpu_resource(gpu_wrapper::gpu_interface *gi) noexcept {
@@ -81,7 +83,7 @@ public:
 
 template <bool is_not_optical>
 class ImageCvter : public GPU_wrapper_wrapper<is_not_optical> {
-public:
+ public:
   using basic_colorset_t = colorset_new<true, is_not_optical>;
   using allowed_colorset_t = colorset_new<false, is_not_optical>;
   using TokiColor_t =
@@ -90,18 +92,26 @@ public:
   using coloridx_t = colorid_t;
 
   // These static member must be implemented by caller
-  static const basic_colorset_t &basic_colorset;
-  static const allowed_colorset_t &allowed_colorset;
+  //  static const basic_colorset_t &basic_colorset;
+  //  static const allowed_colorset_t &allowed_colorset;
 
-protected:
+  ImageCvter(const basic_colorset_t &basic, const allowed_colorset_t &allowed)
+      : basic_colorset{basic}, allowed_colorset{allowed} {}
+
+  ImageCvter(ImageCvter &&) = default;
+
+ protected:
+  const basic_colorset_t &basic_colorset;
+  const allowed_colorset_t &allowed_colorset;
   Eigen::ArrayXX<ARGB> _raw_image;
   ::SCL_convertAlgo algo;
+  bool dither{false};
   std::unordered_map<convert_unit, TokiColor_t, ::hash_cvt_unit> _color_hash;
 
   Eigen::ArrayXX<ARGB> _dithered_image;
   // Eigen::ArrayXX<colorid_t> colorid_matrix;
 
-public:
+ public:
   uiPack ui;
   // SCL_convertAlgo convert_algo{SCL_convertAlgo::RGB_Better};
 
@@ -117,6 +127,8 @@ public:
   inline void clear_color_hash() noexcept { this->_color_hash.clear(); }
 
   inline ::SCL_convertAlgo convert_algo() const noexcept { return this->algo; }
+
+  inline bool is_dither() const noexcept { return this->dither; }
 
   inline int64_t rows() const noexcept { return _raw_image.rows(); }
   inline int64_t cols() const noexcept { return _raw_image.cols(); }
@@ -147,11 +159,12 @@ public:
     }
   }
 
-  bool convert_image(::SCL_convertAlgo __algo, bool dither,
+  bool convert_image(::SCL_convertAlgo __algo, bool _dither,
                      bool try_gpu = false) noexcept {
     if (__algo == ::SCL_convertAlgo::gaCvter) {
       __algo = ::SCL_convertAlgo::RGB_Better;
     }
+    this->dither = _dither;
 
     ui.rangeSet(0, 100, 0);
 
@@ -163,34 +176,46 @@ public:
     }
     ui.rangeSet(0, 100, 50);
 
-    if (dither) {
+    if (this->dither) {
       switch (this->algo) {
-      case ::SCL_convertAlgo::RGB:
-        this->template __impl_dither<::SCL_convertAlgo::RGB>();
-        break;
-      case ::SCL_convertAlgo::RGB_Better:
-        this->template __impl_dither<::SCL_convertAlgo::RGB_Better>();
-        break;
-      case ::SCL_convertAlgo::HSV:
-        this->template __impl_dither<::SCL_convertAlgo::HSV>();
-        break;
-      case ::SCL_convertAlgo::Lab94:
-        this->template __impl_dither<::SCL_convertAlgo::Lab94>();
-        break;
-      case ::SCL_convertAlgo::Lab00:
-        this->template __impl_dither<::SCL_convertAlgo::Lab00>();
-        break;
-      case ::SCL_convertAlgo::XYZ:
-        this->template __impl_dither<::SCL_convertAlgo::XYZ>();
-        break;
+        case ::SCL_convertAlgo::RGB:
+          this->template __impl_dither<::SCL_convertAlgo::RGB>();
+          break;
+        case ::SCL_convertAlgo::RGB_Better:
+          this->template __impl_dither<::SCL_convertAlgo::RGB_Better>();
+          break;
+        case ::SCL_convertAlgo::HSV:
+          this->template __impl_dither<::SCL_convertAlgo::HSV>();
+          break;
+        case ::SCL_convertAlgo::Lab94:
+          this->template __impl_dither<::SCL_convertAlgo::Lab94>();
+          break;
+        case ::SCL_convertAlgo::Lab00:
+          this->template __impl_dither<::SCL_convertAlgo::Lab00>();
+          break;
+        case ::SCL_convertAlgo::XYZ:
+          this->template __impl_dither<::SCL_convertAlgo::XYZ>();
+          break;
 
-      default:
-        abort();
-        return false;
+        default:
+          abort();
+          return false;
       }
     } else {
       this->_dithered_image = this->_raw_image;
     }
+
+    //    for (int64_t idx = 0; idx < this->_dithered_image.size(); idx++) {
+    //      const auto current_color{this->_dithered_image(idx)};
+    //      if (getA(current_color) > 0) {
+    //        continue;
+    //      }
+    //      const convert_unit key{current_color, this->algo};
+    //      if (!this->_color_hash.contains(key)) {
+    //        this->_color_hash.emplace(key, uint8_t{0});
+    //      }
+    //    }
+
     ui.rangeSet(0, 100, 100);
     return true;
     // fill_coloridmat_by_hash(this->colorid_matrix);
@@ -201,10 +226,16 @@ public:
     result.setZero(this->rows(), this->cols());
 
     for (int64_t idx = 0; idx < this->size(); idx++) {
-      auto it = this->_color_hash.find(
-          convert_unit(this->_dithered_image(idx), this->algo));
+      const auto current_color = this->_dithered_image(idx);
+
+      auto it = this->_color_hash.find(convert_unit(current_color, this->algo));
 
       if (it == this->_color_hash.end()) {
+        if (getA(current_color) <= 0) {
+          result(idx) = 0;
+          continue;
+        }
+
         abort();
       }
 
@@ -236,10 +267,13 @@ public:
     assert(r >= 0 && r < this->rows());
     assert(c >= 0 && c < this->cols());
 
-    auto it = this->_color_hash.find(
-        convert_unit{this->_dithered_image(r, c), this->algo});
+    const auto current_color = this->_dithered_image(r, c);
+    auto it = this->_color_hash.find(convert_unit{current_color, this->algo});
     if (it == this->_color_hash.end()) {
-      abort();
+      if (getA(current_color) > 0) {
+        abort();
+      }
+      return 0;
     }
     return it->second.color_id();
   }
@@ -250,10 +284,10 @@ public:
     converted_image(dest.data());
   }
 
-  inline void
-  converted_image(ARGB *const data_dest, int64_t *const rows_dest = nullptr,
-                  int64_t *const cols_dest = nullptr,
-                  const bool is_dest_col_major = true) const noexcept {
+  inline void converted_image(
+      ARGB *const data_dest, int64_t *const rows_dest = nullptr,
+      int64_t *const cols_dest = nullptr,
+      const bool is_dest_col_major = true) const noexcept {
     if (rows_dest != nullptr) {
       *rows_dest = this->rows();
     }
@@ -288,7 +322,7 @@ public:
     }
   }
 
-private:
+ private:
   void add_colors_to_hash() noexcept {
     // this->_color_hash.clear();
 
@@ -298,8 +332,7 @@ private:
       auto it = _color_hash.find(cu);
 
       // this key isn't inserted
-      if (it == _color_hash.end())
-        this->_color_hash.emplace(cu, TokiColor_t());
+      if (it == _color_hash.end()) this->_color_hash.emplace(cu, TokiColor_t());
     }
   }
 
@@ -308,7 +341,6 @@ private:
       this->match_all_TokiColors_cpu();
       return true;
     } else {
-
       if constexpr (gpu_wrapper::have_api) {
         // If converter have gpu resources, compute by gpu
         if (try_gpu && this->have_gpu_resource()) {
@@ -331,14 +363,14 @@ private:
     tasks.clear();
 
     for (auto &pair : _color_hash) {
-      if (!pair.second.is_result_computed())
-        tasks.emplace_back(&pair);
+      if (!pair.second.is_result_computed()) tasks.emplace_back(&pair);
     }
     const size_t taskCount = tasks.size();
 
 #pragma omp parallel for schedule(dynamic)
     for (int taskIdx = 0; taskIdx < (int)taskCount; taskIdx++) {
-      tasks[taskIdx]->second.compute(tasks[taskIdx]->first);
+      tasks[taskIdx]->second.compute(tasks[taskIdx]->first,
+                                     this->allowed_colorset);
     }
     // #warning we should parallelize here
     /*
@@ -373,7 +405,8 @@ private:
     }
   }
 
-  template <typename = void> bool match_all_TokiColors_gpu() noexcept {
+  template <typename = void>
+  bool match_all_TokiColors_gpu() noexcept {
     static_assert(!is_not_optical,
                   "GPU boosting is only avaliable for VisualCraftL.");
     // static_assert(gpu_wrapper::have_api, "No avaliable GPU api.");
@@ -390,7 +423,7 @@ private:
     for (auto &pair : this->_color_hash) {
       if (!pair.second.is_result_computed()) {
         if ((pair.first._ARGB & 0xFF'00'00'00) == 0) {
-          pair.second.compute(pair.first);
+          pair.second.compute(pair.first, this->allowed_colorset);
         } else {
           tasks.emplace_back(&pair);
         }
@@ -412,7 +445,6 @@ private:
     if (gpu_task_count > 0) {
       std::vector<std::array<float, 3>> task_colors(gpu_task_count);
       for (size_t tid = 0; tid < gpu_task_count; tid++) {
-
         if (tasks[tid]->first.algo != algo) {
           return false;
         }
@@ -434,35 +466,35 @@ private:
 
     std::array<const float *, 3> colorset_ptrs{nullptr, nullptr, nullptr};
     switch (algo) {
-    case SCL_convertAlgo::RGB:
-    case SCL_convertAlgo::RGB_Better:
-      colorset_ptrs = {TokiColor_t::Allowed->rgb_data(0),
-                       TokiColor_t::Allowed->rgb_data(1),
-                       TokiColor_t::Allowed->rgb_data(2)};
-      break;
-    case SCL_convertAlgo::HSV:
-      colorset_ptrs = {TokiColor_t::Allowed->hsv_data(0),
-                       TokiColor_t::Allowed->hsv_data(1),
-                       TokiColor_t::Allowed->hsv_data(2)};
-      break;
-    case SCL_convertAlgo::Lab94:
-    case SCL_convertAlgo::Lab00:
-      colorset_ptrs = {TokiColor_t::Allowed->lab_data(0),
-                       TokiColor_t::Allowed->lab_data(1),
-                       TokiColor_t::Allowed->lab_data(2)};
-      break;
-    case SCL_convertAlgo::XYZ:
-      colorset_ptrs = {TokiColor_t::Allowed->xyz_data(0),
-                       TokiColor_t::Allowed->xyz_data(1),
-                       TokiColor_t::Allowed->xyz_data(2)};
-      break;
-    default:
-      abort();
+      case SCL_convertAlgo::RGB:
+      case SCL_convertAlgo::RGB_Better:
+        colorset_ptrs = {this->allowed_colorset.rgb_data(0),
+                         this->allowed_colorset.rgb_data(1),
+                         this->allowed_colorset.rgb_data(2)};
+        break;
+      case SCL_convertAlgo::HSV:
+        colorset_ptrs = {this->allowed_colorset.hsv_data(0),
+                         this->allowed_colorset.hsv_data(1),
+                         this->allowed_colorset.hsv_data(2)};
+        break;
+      case SCL_convertAlgo::Lab94:
+      case SCL_convertAlgo::Lab00:
+        colorset_ptrs = {this->allowed_colorset.lab_data(0),
+                         this->allowed_colorset.lab_data(1),
+                         this->allowed_colorset.lab_data(2)};
+        break;
+      case SCL_convertAlgo::XYZ:
+        colorset_ptrs = {this->allowed_colorset.xyz_data(0),
+                         this->allowed_colorset.xyz_data(1),
+                         this->allowed_colorset.xyz_data(2)};
+        break;
+      default:
+        abort();
     }
 
     if (gpu_task_count > 0) {
       // set colorset for ocl
-      this->gpu->set_colorset_v(TokiColor_t::Allowed->color_count(),
+      this->gpu->set_colorset_v(this->allowed_colorset.color_count(),
                                 colorset_ptrs);
       if (!this->gpu->ok_v()) {
         return false;
@@ -476,7 +508,7 @@ private:
     // compute rest tasks on cpu
     for (uint64_t ctid = 0; ctid < cpu_task_count; ctid++) {
       const uint64_t tid = gpu_task_count + ctid;
-      tasks[tid]->second.compute(tasks[tid]->first);
+      tasks[tid]->second.compute(tasks[tid]->first, this->allowed_colorset);
     }
 
     if (gpu_task_count > 0) {
@@ -491,11 +523,11 @@ private:
         TokiColor_t &tc = tasks[tid]->second;
 
         const uint16_t tempidx = this->gpu->result_idx_v()[tid];
-        if (tempidx >= TokiColor_t::Allowed->color_count()) {
+        if (tempidx >= this->allowed_colorset.color_count()) {
           abort();
         }
 
-        tc.set_gpu_result(TokiColor_t::Allowed->color_id(tempidx),
+        tc.set_gpu_result(this->allowed_colorset.color_id(tempidx),
                           this->gpu->result_diff_v()[tid]);
       }
 
@@ -505,24 +537,25 @@ private:
   template <SCL_convertAlgo algo>
   inline static ARGB ColorCvt(float c0, float c1, float c2) noexcept {
     switch (algo) {
-    case ::SCL_convertAlgo::RGB:
-    case ::SCL_convertAlgo::RGB_Better:
-    case ::SCL_convertAlgo::gaCvter:
-      return RGB2ARGB(c0, c1, c2);
-    case ::SCL_convertAlgo::HSV:
-      return HSV2ARGB(c0, c1, c2);
-    case ::SCL_convertAlgo::Lab00:
-    case ::SCL_convertAlgo::Lab94:
-      return Lab2ARGB(c0, c1, c2);
-    case ::SCL_convertAlgo::XYZ:
-      return XYZ2ARGB(c0, c1, c2);
+      case ::SCL_convertAlgo::RGB:
+      case ::SCL_convertAlgo::RGB_Better:
+      case ::SCL_convertAlgo::gaCvter:
+        return RGB2ARGB(c0, c1, c2);
+      case ::SCL_convertAlgo::HSV:
+        return HSV2ARGB(c0, c1, c2);
+      case ::SCL_convertAlgo::Lab00:
+      case ::SCL_convertAlgo::Lab94:
+        return Lab2ARGB(c0, c1, c2);
+      case ::SCL_convertAlgo::XYZ:
+        return XYZ2ARGB(c0, c1, c2);
     }
     // unreachable
     abort();
     return 0;
   }
 
-  template <SCL_convertAlgo cvt_algo> void __impl_dither() noexcept {
+  template <SCL_convertAlgo cvt_algo>
+  void __impl_dither() noexcept {
     std::array<Eigen::ArrayXXf, 3> dither_c3;
     for (auto &i : dither_c3) {
       i.setZero(this->rows() + 2, this->cols() + 2);
@@ -551,7 +584,7 @@ private:
     for (int64_t row = 0; row < this->rows(); row++) {
       if (is_dir_LR)
         for (int64_t col = 0; col < this->cols(); col++) {
-          if (::getA(this->_raw_image(row, col) <= 0)) {
+          if (::getA(this->_raw_image(row, col)) <= 0) {
             continue;
           }
 
@@ -567,7 +600,7 @@ private:
             convert_unit cu(current_argb, this->algo);
             auto ret = this->_color_hash.emplace(cu, TokiColor_t());
             it_to_old_color = ret.first;
-            it_to_old_color->second.compute(cu);
+            it_to_old_color->second.compute(cu, this->allowed_colorset);
             // inserted_count++;
           }
 
@@ -592,7 +625,7 @@ private:
         }
       else
         for (int64_t col = this->cols() - 1; col >= 0; col--) {
-          if (::getA(this->_raw_image(row, col) <= 0)) {
+          if (::getA(this->_raw_image(row, col)) <= 0) {
             continue;
           }
 
@@ -606,7 +639,7 @@ private:
           if (it_to_old_color == this->_color_hash.end()) {
             auto ret = this->_color_hash.emplace(cu, TokiColor_t());
             it_to_old_color = ret.first;
-            it_to_old_color->second.compute(cu);
+            it_to_old_color->second.compute(cu, this->allowed_colorset);
             // inserted_count++;
           }
 
@@ -636,8 +669,25 @@ private:
     }
     return;
   }
+
+ public:
+  [[deprecated]] uint64_t task_hash() const noexcept {
+    return this->task_hash(this->algo, this->dither);
+  }
+
+  uint64_t task_hash(SCL_convertAlgo a, bool d) const noexcept {
+    const auto &img = this->_raw_image;
+    return std::hash<std::string_view>()(
+
+               std::string_view{(const char *)img.data(),
+                                img.size() * sizeof(uint32_t)}) ^
+           std::hash<char>()((char)a) ^ std::hash<bool>()(d);
+  }
 };
 
-} // namespace libImageCvt
+std::vector<uint8_t> hash_of_image(
+    Eigen::Map<const Eigen::ArrayXX<ARGB>> img) noexcept;
 
-#endif // COLORMANIP_IMAGECONVERT_HPP
+}  // namespace libImageCvt
+
+#endif  // COLORMANIP_IMAGECONVERT_HPP
