@@ -30,6 +30,62 @@ This file is part of SlopeCraft.
 namespace MCDataVersion {
 
 enum class MCDataVersion_t : int {
+  Java_1_21_1 = 3955,
+  Java_1_21_1__rc1 = 3954,
+  Java_1_21 = 3953,
+  Java_1_21__rc1 = 3952,
+  Java_1_21__pre4 = 3951,
+  Java_1_21__pre3 = 3950,
+  Java_1_21__pre2 = 3949,
+  Java_1_21__pre1 = 3948,
+  Snapshot_24w21b = 3947,
+  Snapshot_24w21a = 3946,
+  Snapshot_24w20a = 3944,
+  Snapshot_24w19b = 3942,
+  Snapshot_24w19a = 3941,
+  Snapshot_24w18a = 3940,
+  Java_1_20_6 = 3839,
+  Java_1_20_6__rc1 = 3838,
+  Java_1_20_5 = 3837,
+  Snapshot_23w51b = 3802,
+  Snapshot_23w51a = 3801,
+  Java_1_20_4 = 3700,
+  Java_1_20_4_rc1 = 3699,
+  Java_1_20_3 = 3698,
+  Java_1_20_3__rc2 = 3577,
+  Java_1_20_3__rc1 = 3576,
+  Java_1_20_2_pre4 = 3575,
+  Java_1_20_2_pre3 = 3574,
+  Java_1_20_2_pre2 = 3573,
+  Java_1_20_2_pre1 = 3572,
+  Snapshot_23w35a = 3571,
+  Snapshot_23w33a = 3570,
+  Snapshot_23w32a = 3569,
+  Snapshot_23w31a = 3567,
+  Java_1_20_1 = 3465,
+  Java_1_20_1__rc1 = 3464,
+  Java_1_20 = 3463,
+  Java_1_20__rc1 = 3462,
+  Java_1_20__pre7 = 3461,
+  Java_1_20__pre6 = 3460,
+  Java_1_20__pre5 = 3458,
+  Java_1_20__pre4 = 3457,
+  Java_1_20__pre3 = 3456,
+  Java_1_20__pre2 = 3455,
+  Java_1_20__pre1 = 3454,
+  Snapshot_23w_18a = 3453,
+  Snapshot_23w_17a = 3452,
+  Snapshot_23w_16a = 3449,
+  Snapshot_23w_14a = 3445,
+  Snapshot_23w_13a = 3442,
+  Snapshot_23w_12a = 3442,
+  Java_1_19_4 = 3337,
+  Java_1_19_4__rc3 = 3336,
+  Java_1_19_4__rc2 = 3335,
+  Java_1_19_4__rc1 = 3334,
+  Java_1_19_4__pre4 = 3333,
+  Java_1_19_4__pre3 = 3332,
+  Java_1_19_4__pre2 = 3331,
   Java_1_19_4__pre1 = 3330,
   Snapshot_23w07a = 3329,
   Snapshot_23w06a = 3326,
@@ -515,5 +571,5 @@ inline bool is_data_version_suitable(SCL_gameVersion v,
   return true;
 }
 
-} // namespace MCDataVersion
-#endif // SLOPECRAFT_UTILITIES_MCDATAVERSION_H
+}  // namespace MCDataVersion
+#endif  // SLOPECRAFT_UTILITIES_MCDATAVERSION_H
