@@ -25,12 +25,16 @@ This file is part of SlopeCraft.
 
 #include <ColorManip/ColorManip.h>
 #include <ColorManip/imageConvert.hpp>
-
+#include <cereal/cereal.hpp>
+#include <cereal/access.hpp>
+#include <cereal/types/unordered_map.hpp>
 #include <memory>
+#include <exception>
 
 namespace GACvter {
 class GAConverter;
-}
+void delete_GA_converter(GAConverter *) noexcept;
+}  // namespace GACvter
 
 namespace heu {
 struct GAOption;
@@ -39,17 +43,22 @@ struct GAOption;
 namespace libMapImageCvt {
 
 class MapImageCvter : public ::libImageCvt::ImageCvter<true> {
-private:
-  std::unique_ptr<GACvter::GAConverter> gacvter;
+ private:
+  using deleter_t = decltype([](GACvter::GAConverter *g) {
+    GACvter::delete_GA_converter(g);
+  });
+  std::unique_ptr<GACvter::GAConverter, deleter_t> gacvter;
 
-public:
+ public:
   using Base_t = ::libImageCvt::ImageCvter<true>;
 
   using TokiColor_t = typename Base_t::TokiColor_t;
 
   static constexpr int size_of_tokicolor = sizeof(TokiColor_t);
 
-  explicit MapImageCvter();
+  MapImageCvter(const Base_t::basic_colorset_t &basic,
+                const Base_t::allowed_colorset_t &allowed);
+  MapImageCvter(MapImageCvter &&) = default;
 
   ~MapImageCvter() = default;
 
@@ -75,23 +84,111 @@ public:
     return result;
   }
 
-  void col_TokiColor_ptrs(int64_t col,
-                          const TokiColor_t **const dest) const noexcept {
-    if (dest == nullptr) {
-      return;
-    }
-
+  void col_TokiColor_ptrs(
+      int64_t col, std::vector<const TokiColor_t *> &dest) const noexcept {
     if (col < 0 || col > this->cols()) {
       return;
     }
-
+    dest.clear();
     for (int64_t r = 0; r < this->rows(); r++) {
       auto it = this->_color_hash.find(
           convert_unit(this->_dithered_image(r, col), this->algo));
-      dest[r] = &it->second;
+      dest.emplace_back(&it->second);
     }
   }
+  
+  // temp is a temporary container to pass ownership
+  void load_from_itermediate(MapImageCvter &&temp) noexcept {
+    this->_raw_image = std::move(temp._raw_image);
+    this->algo = temp.algo;
+    this->_dithered_image = std::move(temp._dithered_image);
+
+    assert(this->_raw_image.rows() == this->_dithered_image.rows());
+    assert(this->_raw_image.cols() == this->_dithered_image.cols());
+    if (this->_color_hash.empty()) {
+      this->_color_hash = std::move(temp._color_hash);
+    } else {
+      temp._color_hash.merge(this->_color_hash);
+      this->_color_hash = std::move(temp._color_hash);
+    }
+  }
+
+ private:
+  friend class cereal::access;
+  template <class archive>
+  void save(archive &ar) const {
+    assert(this->_raw_image.rows() == this->_dithered_image.rows());
+    assert(this->_raw_image.cols() == this->_dithered_image.cols());
+    ar(this->_raw_image);
+    ar(this->algo);
+    ar(this->dither);
+    // ar(this->_color_hash);
+    ar(this->_dithered_image);
+    // save required colorset
+    {
+      std::unordered_set<uint32_t> colors_dithered_img;
+      colors_dithered_img.reserve(this->_dithered_image.size());
+      for (int64_t i = 0; i < this->_dithered_image.size(); i++) {
+        colors_dithered_img.emplace(this->_dithered_image(i));
+        colors_dithered_img.emplace(this->_raw_image(i));
+      }
+
+      const size_t size_colorset = colors_dithered_img.size();
+      ar(size_colorset);
+      for (uint32_t color : colors_dithered_img) {
+        auto it =
+            this->color_hash().find(convert_unit{color, this->convert_algo()});
+        if (it == this->color_hash().end()) {
+          assert(getA(color) <= 0);
+          continue;
+        }
+
+        ar(it->first, it->second);
+      }
+    }
+  }
+
+  template <class archive>
+  void load(archive &ar) {
+    ar(this->_raw_image);
+    ar(this->algo);
+    ar(this->dither);
+    // ar(this->_color_hash);
+    ar(this->_dithered_image);
+
+    assert(this->_raw_image.rows() == this->_dithered_image.rows());
+    assert(this->_raw_image.cols() == this->_dithered_image.cols());
+
+    {
+      size_t size_colorset{0};
+      ar(size_colorset);
+      for (size_t idx = 0; idx < size_colorset; idx++) {
+        convert_unit key;
+        TokiColor_t val;
+        ar(key, val);
+
+        this->_color_hash.emplace(key, val);
+      }
+
+      for (int64_t i = 0; i < this->_dithered_image.size(); i++) {
+        auto it = this->_color_hash.find(
+            convert_unit{this->_dithered_image(i), this->convert_algo()});
+        if (it == this->_color_hash.end()) {
+          throw std::runtime_error{
+              "One or more colors not found in cached colorhash"};
+        }
+      }
+    }
+  }
+
+ public:
+  bool save_cache(const char *filename) const noexcept;
+  bool examine_cache(const char *filename, uint64_t expected_task_hash,
+                     MapImageCvter *itermediate = nullptr) const noexcept;
+  [[nodiscard]] bool load_cache(const char *filename,
+                                uint64_t expected_task_hash) noexcept;
+  bool load_cache(const char *filename) noexcept;
 };
 
-} // namespace libMapImageCvt
-#endif // SCL_MAPIMAGECVTER_MAPIMAGECVTER_H
+}  // namespace libMapImageCvt
+#endif  // SCL_MAPIMAGECVTER_MAPIMAGECVTER_H
