@@ -20,13 +20,16 @@ This file is part of SlopeCraft.
     bilibili:https://space.bilibili.com/351429231
 */
 
-#include "Schem.h"
-
+#include <Eigen/Dense>
+#include <Eigen/StdVector>
 #include <memory.h>
-
+#include <ranges>
 #include <ctime>
 #include <filesystem>
 #include <iostream>
+#include <fmt/format.h>
+
+#include "Schem.h"
 
 #include "../NBTWriter/NBTWriter.h"
 #include "bit_shrink.h"
@@ -38,6 +41,36 @@ WorldEditSchem_info::WorldEditSchem_info() : date(std::time(nullptr) * 1000) {}
 litematic_info::litematic_info()
     : time_created(std::time(nullptr) * 1000),
       time_modified(std::time(nullptr) * 1000) {}
+
+void Schem::resize(int64_t x, int64_t y, int64_t z) {
+  if (x < 0 || y < 0 || z < 0) {
+    return;
+  }
+  this->xzy.resize(x, z, y);
+}
+
+std::string Schem::check_size() const noexcept {
+  return check_size(this->x_range(), this->y_range(), this->z_range());
+}
+
+std::string Schem::check_size(int64_t x, int64_t y, int64_t z) noexcept {
+  if (x < 0 || y < 0 || z < 0) {
+    return fmt::format(
+        "Size in one or more dimensions is negative. x= {}, y= {}, z= {}", x, y,
+        z);
+  }
+  return {};
+}
+
+void Schem::stat_blocks(std::vector<size_t> &dest) const noexcept {
+  dest.resize(this->palette_size());
+  std::fill(dest.begin(), dest.end(), 0);
+
+  for (ele_t block_index : *this) {
+    assert(block_index < this->palette_size());
+    dest[block_index] += 1;
+  }
+}
 
 void Schem::set_block_id(const char *const *const block_ids,
                          const int num) noexcept {
@@ -51,6 +84,13 @@ void Schem::set_block_id(const char *const *const block_ids,
 
   for (int idx = 0; idx < num; idx++) {
     this->block_id_list[idx] = block_ids[idx];
+  }
+}
+
+void Schem::set_block_id(std::span<std::string_view> id) noexcept {
+  this->block_id_list.resize(id.size());
+  for (size_t idx = 0; idx < id.size(); idx++) {
+    this->block_id_list[idx] = id[idx];
   }
 }
 
@@ -100,7 +140,105 @@ bool Schem::have_invalid_block(
   return false;
 }
 
-enum class __mushroom_type : uint8_t {
+Schem Schem::slice_no_check(std::span<const std::pair<int64_t, int64_t>, 3>
+                                xyz_index_range) const noexcept {
+  Schem ret;
+  ret.block_id_list = this->block_id_list;
+  ret.MC_major_ver = this->MC_major_ver;
+  ret.MC_data_ver = this->MC_data_ver;
+
+  ret.entities.reserve(this->entities.size());
+  Eigen::Array3d pos_min, pos_max;
+  for (size_t dim = 0; dim < 3; dim++) {
+    pos_min[dim] = xyz_index_range[dim].first;
+    pos_max[dim] = xyz_index_range[dim].second;
+  }
+  {
+    Eigen::array<int64_t, 3> start, extent;
+    start[0] = xyz_index_range[0].first;
+    start[1] = xyz_index_range[2].first;
+    start[2] = xyz_index_range[1].first;
+    extent[0] = xyz_index_range[0].second - xyz_index_range[0].first;
+    extent[1] = xyz_index_range[2].second - xyz_index_range[2].first;
+    extent[2] = xyz_index_range[1].second - xyz_index_range[1].first;
+    ret.xzy = this->xzy.slice(start, extent);
+  }
+
+  for (auto &entity : this->entities) {
+    const auto pos_std = entity->position();
+    const Eigen::Array3d pos{pos_std[0], pos_std[1], pos_std[2]};
+    if ((pos >= pos_min and pos < pos_max).all()) {
+      ret.entities.emplace_back(entity->clone());
+    }
+  }
+
+  return ret;
+}
+
+tl::expected<boost::multi_array<Schem::schem_slice<Schem>, 3>, std::string>
+Schem::split_by_block_size(
+    std::span<const uint64_t> x_block_length,
+    std::span<const uint64_t> y_block_length,
+    std::span<const uint64_t> z_block_length) const noexcept {
+  // check input block length and compute index ranges
+  std::array<std::vector<std::pair<int64_t, int64_t>>, 3> xyz_block_index_pairs;
+  {
+    std::array<const std::span<const uint64_t>, 3> xyz_block_len{
+        x_block_length, y_block_length, z_block_length};
+    std::array<uint64_t, 3> block_len_sum{0, 0, 0};
+    std::array<int64_t, 3> shape{this->x_range(), this->y_range(),
+                                 this->z_range()};
+    for (size_t dim = 0; dim < 3; dim++) {
+      int64_t cur_block_start_index = 0;
+      for (auto [blk_idx, block_len] :
+           xyz_block_len[dim] | std::views::enumerate) {
+        if (block_len <= 0) {
+          return tl::make_unexpected(fmt::format(
+              "Found non-positive block length in dim = {}, block index = {}",
+              dim, blk_idx));
+        }
+        block_len_sum[dim] += block_len;
+        const int64_t cur_block_end_index =
+            static_cast<int64_t>(block_len_sum[dim]);
+        assert(cur_block_end_index > cur_block_start_index);
+        xyz_block_index_pairs[dim].emplace_back(cur_block_start_index,
+                                                cur_block_end_index);
+        cur_block_start_index = cur_block_end_index;
+      }
+      if (block_len_sum[dim] not_eq shape[dim]) {
+        return tl::make_unexpected(fmt::format(
+            "The sum of block length of dim {} is {}, which is not identical "
+            "to shape on this dim({}), ",
+            dim, block_len_sum[dim], shape[dim]));
+      }
+    }
+  }
+
+  boost::multi_array<Schem::schem_slice<Schem>, 3> ret{
+      boost::extents[xyz_block_index_pairs[0].size()]
+                    [xyz_block_index_pairs[1].size()]
+                    [xyz_block_index_pairs[2].size()]};
+
+  for (auto [x_idx, x_range] :
+       xyz_block_index_pairs[0] | std::views::enumerate) {
+    for (auto [y_idx, y_range] :
+         xyz_block_index_pairs[1] | std::views::enumerate) {
+      for (auto [z_idx, z_range] :
+           xyz_block_index_pairs[2] | std::views::enumerate) {
+        const Eigen::Array<int64_t, 3, 1> offset{x_range.first, y_range.first,
+                                                 z_range.first};
+        auto &dest = ret[x_idx][y_idx][z_idx];
+        dest.offset = offset;
+        dest.content =
+            this->slice_no_check(std::array<std::pair<int64_t, int64_t>, 3>{
+                x_range, y_range, z_range});
+      }
+    }
+  }
+  return ret;
+}
+
+enum class mushroom_type : uint8_t {
   not_mushroom = 0,
   red_mushroom = 1,
   brown_mushroom = 2,
@@ -154,7 +292,7 @@ void Schem::process_mushroom_states() noexcept {
       u6_to_ele_stem[u6] = this->block_id_list.size() - 1;
     }
   }
-  std::vector<__mushroom_type> is_mushroom_LUT;
+  std::vector<mushroom_type> is_mushroom_LUT;
   is_mushroom_LUT.resize(this->block_id_list.size());
 
   for (int blockidx = 0; blockidx < int(this->block_id_list.size());
@@ -162,18 +300,18 @@ void Schem::process_mushroom_states() noexcept {
     const auto &block_id = this->block_id_list[blockidx];
     const auto pure_id = ::to_pure_block_id(block_id);
     if (pure_id == id_red) {
-      is_mushroom_LUT[blockidx] = __mushroom_type::red_mushroom;
+      is_mushroom_LUT[blockidx] = mushroom_type::red_mushroom;
       continue;
     }
     if (pure_id == id_brown) {
-      is_mushroom_LUT[blockidx] = __mushroom_type::brown_mushroom;
+      is_mushroom_LUT[blockidx] = mushroom_type::brown_mushroom;
       continue;
     }
     if (pure_id == id_stem) {
-      is_mushroom_LUT[blockidx] = __mushroom_type::mushroom_stem;
+      is_mushroom_LUT[blockidx] = mushroom_type::mushroom_stem;
       continue;
     }
-    is_mushroom_LUT[blockidx] = __mushroom_type::not_mushroom;
+    is_mushroom_LUT[blockidx] = mushroom_type::not_mushroom;
   }
 
   // fix the correct state
@@ -181,9 +319,9 @@ void Schem::process_mushroom_states() noexcept {
     for (int64_t z = 0; z < z_range(); z++) {
       for (int64_t x = 0; x < x_range(); x++) {
         // if current block is not mushroom, continue
-        const __mushroom_type current_mushroom_type =
+        const mushroom_type current_mushroom_type =
             is_mushroom_LUT[this->operator()(x, y, z)];
-        if (current_mushroom_type == __mushroom_type::not_mushroom) {
+        if (current_mushroom_type == mushroom_type::not_mushroom) {
           continue;
         }
 
@@ -191,42 +329,42 @@ void Schem::process_mushroom_states() noexcept {
         // match the correct side
         if (x + 1 < x_range()) {
           const ele_t ele_of_side = this->operator()(x + 1, y, z);
-          if (is_mushroom_LUT[ele_of_side] != __mushroom_type::not_mushroom) {
+          if (is_mushroom_LUT[ele_of_side] != mushroom_type::not_mushroom) {
             side.east() = false;
           }
         }
 
         if (x - 1 >= 0) {
           const ele_t ele_of_side = this->operator()(x - 1, y, z);
-          if (is_mushroom_LUT[ele_of_side] != __mushroom_type::not_mushroom) {
+          if (is_mushroom_LUT[ele_of_side] != mushroom_type::not_mushroom) {
             side.west() = false;
           }
         }
 
         if (y + 1 < y_range()) {
           const ele_t ele_of_side = this->operator()(x, y + 1, z);
-          if (is_mushroom_LUT[ele_of_side] != __mushroom_type::not_mushroom) {
+          if (is_mushroom_LUT[ele_of_side] != mushroom_type::not_mushroom) {
             side.up() = false;
           }
         }
 
         if (y - 1 >= 0) {
           const ele_t ele_of_side = this->operator()(x, y - 1, z);
-          if (is_mushroom_LUT[ele_of_side] != __mushroom_type::not_mushroom) {
+          if (is_mushroom_LUT[ele_of_side] != mushroom_type::not_mushroom) {
             side.down() = false;
           }
         }
 
         if (z + 1 < z_range()) {
           const ele_t ele_of_side = this->operator()(x, y, z + 1);
-          if (is_mushroom_LUT[ele_of_side] != __mushroom_type::not_mushroom) {
+          if (is_mushroom_LUT[ele_of_side] != mushroom_type::not_mushroom) {
             side.south() = false;
           }
         }
 
         if (z - 1 >= 0) {
           const ele_t ele_of_side = this->operator()(x, y, z - 1);
-          if (is_mushroom_LUT[ele_of_side] != __mushroom_type::not_mushroom) {
+          if (is_mushroom_LUT[ele_of_side] != mushroom_type::not_mushroom) {
             side.north() = false;
           }
         }
@@ -234,10 +372,10 @@ void Schem::process_mushroom_states() noexcept {
         // write in the correct value of ele_t
         ele_t corrected_ele = invalid_ele_t;
         switch (current_mushroom_type) {
-          case __mushroom_type::brown_mushroom:
+          case mushroom_type::brown_mushroom:
             corrected_ele = u6_to_ele_brown[side.u6()];
             break;
-          case __mushroom_type::red_mushroom:
+          case mushroom_type::red_mushroom:
             corrected_ele = u6_to_ele_red[side.u6()];
             break;
           default:
@@ -252,49 +390,42 @@ void Schem::process_mushroom_states() noexcept {
   return;
 }
 
-bool Schem::export_litematic(std::string_view filename,
-                             const litematic_info &info,
-                             SCL_errorFlag *const error_flag,
-                             std::string *const error_str) const noexcept {
-  if (std::filesystem::path(filename).extension() != ".litematic") {
+tl::expected<void, std::pair<SCL_errorFlag, std::string>> Schem::pre_check(
+    std::string_view filename, std::string_view extension) const noexcept {
+  if (std::filesystem::path(filename).extension() != extension) {
     // wrong extension
-
-    if (error_flag != nullptr) {
-      *error_flag = SCL_errorFlag::EXPORT_SCHEM_WRONG_EXTENSION;
-    }
-    if (error_str != nullptr) {
-      *error_str = "The filename externsion must be \".litematic\".";
-    }
-    return false;
+    return tl::make_unexpected(std::make_pair(
+        SCL_errorFlag::EXPORT_SCHEM_WRONG_EXTENSION,
+        fmt::format("The filename extension must be \"{}\".", extension)));
   }
   // check for invalid blocks
   {
     std::array<int64_t, 3> pos;
     if (this->have_invalid_block(&pos[0], &pos[1], &pos[2])) {
-      if (error_flag != nullptr) {
-        *error_flag = SCL_errorFlag::EXPORT_SCHEM_HAS_INVALID_BLOCKS;
-      }
-      if (error_str != nullptr) {
-        *error_str = "The first invalid block is at x=";
-        *error_str += std::to_string(pos[0]) + ", y=";
-        *error_str += std::to_string(pos[1]) + ", z=";
-        *error_str += std::to_string(pos[2]);
-      }
-      return false;
+      return tl::make_unexpected(std::make_pair(
+          SCL_errorFlag::EXPORT_SCHEM_HAS_INVALID_BLOCKS,
+          fmt::format("The first invalid block is at x={}, y={}, z={}", pos[0],
+                      pos[1], pos[2])));
     }
   }
+  return {};
+}
 
+tl::expected<void, std::pair<SCL_errorFlag, std::string>>
+Schem::export_litematic(std::string_view filename,
+                        const litematic_info &info) const noexcept {
+  {
+    auto res = this->pre_check(filename, ".litematic");
+    if (not res) {
+      return res;
+    }
+  }
   NBT::NBTWriter<true> lite;
 
   if (!lite.open(filename.data())) {
-    if (error_flag != nullptr) {
-      *error_flag = SCL_errorFlag::EXPORT_SCHEM_FAILED_TO_CREATE_FILE;
-    }
-    if (error_str != nullptr) {
-      *error_str = "Failed to open file : ";
-      *error_str += filename;
-    }
-    return false;
+    return tl::make_unexpected(
+        std::make_pair(SCL_errorFlag::EXPORT_SCHEM_FAILED_TO_CREATE_FILE,
+                       fmt::format("Failed to open file: {}", filename)));
   }
 
   lite.writeCompound("Metadata");
@@ -345,14 +476,13 @@ bool Schem::export_litematic(std::string_view filename,
 
       // progressRangeSet(wind, 0, 100 + Build.size(), 100);
 
-      // reportWorkingStatue(wind, workStatues::writingBlockPalette);
+      // reportWorkingStatue(wind, workStatus::writingBlockPalette);
       // write block palette
       lite.writeListHead("BlockStatePalette", NBT::Compound,
                          this->palette_size());
       {
         std::string pure_block_id;
         pure_block_id.reserve(1024);
-        memset(pure_block_id.data(), 0, pure_block_id.capacity());
         std::vector<std::pair<std::string, std::string>> properties;
         properties.reserve(64);
 
@@ -376,10 +506,8 @@ bool Schem::export_litematic(std::string_view filename,
         }
       }
 
-      lite.writeListHead("Entities", NBT::Compound, 0);
       lite.writeListHead("PendingBlockTicks", NBT::Compound, 0);
-      lite.writeListHead("PendingFluidTiccks", NBT::Compound, 0);
-      lite.writeListHead("TileEntities", NBT::Compound, 0);
+      lite.writeListHead("PendingFluidTicks", NBT::Compound, 0);
 
       // write 3D
       std::vector<uint64_t> shrinked;
@@ -394,6 +522,21 @@ bool Schem::export_litematic(std::string_view filename,
         }
       }
       // progressAdd(wind, size3D[0]);
+
+      lite.writeListHead("Entities", NBT::tagType::Compound,
+                         this->entities.size());
+      for (auto &entity : this->entities) {
+        assert(entity);
+        lite.writeCompound();
+        auto res = entity->dump(lite, this->MC_data_ver);
+        if (not res) {
+          lite.endCompound();
+          return tl::make_unexpected(
+              std::make_pair(SCL_errorFlag::EXPORT_SCHEM_HAS_INVALID_ENTITY,
+                             std::move(res.error())));
+        }
+        lite.endCompound();
+      }
     }
     lite.endCompound();  // end current region
   }
@@ -411,64 +554,31 @@ bool Schem::export_litematic(std::string_view filename,
     case ::SCL_gameVersion::MC17:
     case ::SCL_gameVersion::MC18:
     case ::SCL_gameVersion::MC19:
+    case ::SCL_gameVersion::MC20:
+    case ::SCL_gameVersion::MC21:
       lite.writeInt("MinecraftDataVersion", (int)this->MC_version_number());
       lite.writeInt("Version", 5);
       break;
     default:
-      std::cerr << "Wrong game version!" << std::endl;
       lite.close();
-      if (error_flag != nullptr) {
-        *error_flag = SCL_errorFlag::UNKNOWN_MAJOR_GAME_VERSION;
-      }
-      if (error_str != nullptr) {
-        *error_str =
-            "Unknown major game version! Only 1.12 to 1.19 is "
-            "supported, but given value " +
-            std::to_string((int)this->MC_major_ver);
-      }
-      return false;
+      return tl::make_unexpected(std::make_pair(
+          SCL_errorFlag::UNKNOWN_MAJOR_GAME_VERSION,
+          fmt::format("Unknown major game version! Only 1.12 to 1.19 is "
+                      "supported, but given value {}",
+                      int(this->MC_major_ver))));
   }
   lite.close();
 
-  if (error_flag != nullptr) {
-    *error_flag = SCL_errorFlag::NO_ERROR_OCCUR;
-  }
-  if (error_str != nullptr) {
-    *error_str = "";
-  }
-  return true;
+  return {};
 }
 
-bool Schem::export_structure(std::string_view filename,
-                             const bool is_air_structure_void,
-                             SCL_errorFlag *const error_flag,
-                             std::string *const error_str) const noexcept {
-  if (std::filesystem::path(filename).extension() != ".nbt") {
-    // wrong extension
-
-    if (error_flag != nullptr) {
-      *error_flag = SCL_errorFlag::EXPORT_SCHEM_WRONG_EXTENSION;
-    }
-    if (error_str != nullptr) {
-      *error_str = "The filename externsion must be \".nbt\".";
-    }
-    return false;
-  }
-
-  // check for invalid blocks
+tl::expected<void, std::pair<SCL_errorFlag, std::string>>
+Schem::export_structure(std::string_view filename,
+                        const bool is_air_structure_void) const noexcept {
   {
-    std::array<int64_t, 3> pos;
-    if (this->have_invalid_block(&pos[0], &pos[1], &pos[2])) {
-      if (error_flag != nullptr) {
-        *error_flag = SCL_errorFlag::EXPORT_SCHEM_HAS_INVALID_BLOCKS;
-      }
-      if (error_str != nullptr) {
-        *error_str = "The first invalid block is at x=";
-        *error_str += std::to_string(pos[0]) + ", y=";
-        *error_str += std::to_string(pos[1]) + ", z=";
-        *error_str += std::to_string(pos[2]);
-      }
-      return false;
+    auto res = this->pre_check(filename, ".nbt");
+    if (not res) {
+      return res;
     }
   }
 
@@ -487,42 +597,31 @@ bool Schem::export_structure(std::string_view filename,
                  "minecraft:air in your block palette."
               << std::endl;
 
-    if (error_flag != nullptr) {
-      *error_flag = SCL_errorFlag::EXPORT_SCHEM_STRUCTURE_REQUIRES_AIR;
-    }
-    if (error_str != nullptr) {
-      *error_str =
-          "You assigned is_air_structure_void=false, but there is no "
-          "minecraft:air in your block palette.";
-    }
-    return false;
+    return tl::make_unexpected(std::make_pair(
+        SCL_errorFlag::EXPORT_SCHEM_STRUCTURE_REQUIRES_AIR,
+        "You assigned is_air_structure_void=false, but there is no "
+        "minecraft:air in your block palette."));
   }
 
   /*
-  reportWorkingStatue(wind, workStatues::writingMetaInfo);
+  reportWorkingStatue(wind, workStatus::writingMetaInfo);
   progressRangeSet(wind, 0, 100 + Build.size(), 0);
   */
 
   NBT::NBTWriter<true> file;
   if (!file.open(filename.data())) {
-    if (error_flag != nullptr) {
-      *error_flag = SCL_errorFlag::EXPORT_SCHEM_FAILED_TO_CREATE_FILE;
-    }
-    if (error_str != nullptr) {
-      *error_str = "Failed to open file : ";
-      *error_str += filename;
-    }
-    return false;
+    return tl::make_unexpected(
+        std::make_pair(SCL_errorFlag::EXPORT_SCHEM_FAILED_TO_CREATE_FILE,
+                       fmt::format("Failed to open file {}", filename)));
   }
 
-  file.writeListHead("entities", NBT::Byte, 0);
   file.writeListHead("size", NBT::Int, 3);
   {
     file.writeInt("This should never be shown", x_range());
     file.writeInt("This should never be shown", y_range());
     file.writeInt("This should never be shown", z_range());
   }
-  // reportWorkingStatue(wind, workStatues::writingBlockPalette);
+  // reportWorkingStatue(wind, workStatus::writingBlockPalette);
   file.writeListHead("palette", NBT::Compound, palette_size());
   {
     std::string pure_block_id;
@@ -594,7 +693,41 @@ bool Schem::export_structure(std::string_view filename,
         }
       }
     }
-    // finish writting the whole 3D array
+    // finish writing the whole 3D array
+
+    // write entities
+    file.writeListHead("entities", NBT::tagType::Compound,
+                       this->entities.size());
+    for (auto &entity : this->entities) {
+      file.writeCompound();
+      {
+        file.writeListHead("pos", NBT::tagType::Double, 3);
+        for (double pos : entity->position()) {
+          file.writeDouble("", pos);
+        }
+        assert(file.isInCompound());
+        file.writeListHead("blockPos", NBT::tagType::Int, 3);
+        for (double pos : entity->position()) {
+          file.writeInt("", std::floor(pos));
+        }
+        assert(file.isInCompound());
+        file.writeCompound("nbt");
+        {
+          auto res = entity->dump(file, this->MC_data_ver);
+          if (not res) {
+            file.endCompound();
+            file.close_file();
+            return tl::make_unexpected(
+                std::make_pair(SCL_errorFlag::EXPORT_SCHEM_HAS_INVALID_ENTITY,
+                               std::move(res.error())));
+          }
+        }
+        file.endCompound();
+      }
+      file.endCompound();
+    }
+    // finish writing entities
+    assert(file.isInCompound());
 
     switch (this->MC_major_ver) {
       case ::SCL_gameVersion::MC12:
@@ -605,158 +738,159 @@ bool Schem::export_structure(std::string_view filename,
       case ::SCL_gameVersion::MC17:
       case ::SCL_gameVersion::MC18:
       case ::SCL_gameVersion::MC19:
+      case ::SCL_gameVersion::MC20:
+      case ::SCL_gameVersion::MC21:
         file.writeInt("MinecraftDataVersion", (int)this->MC_data_ver);
         break;
       default:
         std::cerr << "Wrong game version!" << std::endl;
         file.close();
-        if (error_flag != nullptr) {
-          *error_flag = SCL_errorFlag::UNKNOWN_MAJOR_GAME_VERSION;
-        }
-        if (error_str != nullptr) {
-          *error_str =
-              "Unknown major game version! Only 1.12 to 1.19 is "
-              "supported, but given value " +
-              std::to_string((int)this->MC_major_ver);
-        }
-        return false;
+        return tl::make_unexpected(std::make_pair(
+            SCL_errorFlag::UNKNOWN_MAJOR_GAME_VERSION,
+            fmt::format("Unknown major game version! Only 1.12 to 1.21 is "
+                        "supported, but given value {}",
+                        (int)this->MC_major_ver)));
     }
   }
   file.close();
 
-  if (error_flag != nullptr) {
-    *error_flag = SCL_errorFlag::NO_ERROR_OCCUR;
-  }
-  if (error_str != nullptr) {
-    *error_str = "";
-  }
-  return true;
+  return {};
 }
 
-bool Schem::export_WESchem(std::string_view filename,
-                           const WorldEditSchem_info &info,
-                           SCL_errorFlag *const error_flag,
-                           std::string *const error_str) const noexcept {
-  if (std::filesystem::path(filename).extension() != ".schem") {
-    // wrong extension
-
-    if (error_flag != nullptr) {
-      *error_flag = SCL_errorFlag::EXPORT_SCHEM_WRONG_EXTENSION;
-    }
-    if (error_str != nullptr) {
-      *error_str = "The filename externsion must be \".schem\".";
-    }
-    return false;
-  }
-
-  // check for invalid blocks
+tl::expected<void, std::pair<SCL_errorFlag, std::string>> Schem::export_WESchem(
+    std::string_view filename, const WorldEditSchem_info &info) const noexcept {
   {
-    std::array<int64_t, 3> pos;
-    if (this->have_invalid_block(&pos[0], &pos[1], &pos[2])) {
-      if (error_flag != nullptr) {
-        *error_flag = SCL_errorFlag::EXPORT_SCHEM_HAS_INVALID_BLOCKS;
-      }
-      if (error_str != nullptr) {
-        *error_str = "The first invalid block is at x=";
-        *error_str += std::to_string(pos[0]) + ", y=";
-        *error_str += std::to_string(pos[1]) + ", z=";
-        *error_str += std::to_string(pos[2]);
-      }
-      return false;
+    auto res = this->pre_check(filename, ".schem");
+    if (not res) {
+      return res;
     }
   }
 
   if (this->MC_major_ver <= SCL_gameVersion::MC12) {
-    if (error_flag != nullptr) {
-      *error_flag = ::SCL_errorFlag::EXPORT_SCHEM_MC12_NOT_SUPPORTED;
-    }
-    if (error_str != nullptr) {
-      *error_str =
-          "Exporting a schematic as 1.12 WorldEdit .schematic format "
-          "is not supported. Try other tools.";
-    }
-    return false;
+    return tl::make_unexpected(std::make_pair(
+        ::SCL_errorFlag::EXPORT_SCHEM_MC12_NOT_SUPPORTED,
+        "Exporting a schematic as 1.12 WorldEdit .schematic format "
+        "is not supported. Try other tools."));
   }
 
   NBT::NBTWriter<true> file;
 
-  if (!file.open(filename.data())) {
-    if (error_flag != nullptr) {
-      *error_flag = SCL_errorFlag::EXPORT_SCHEM_FAILED_TO_CREATE_FILE;
-    }
-    if (error_str != nullptr) {
-      *error_str = "Failed to open file : ";
-      *error_str += filename;
-    }
-    return false;
+  if (not file.open(filename.data())) {
+    return tl::make_unexpected(
+        std::make_pair(SCL_errorFlag::EXPORT_SCHEM_FAILED_TO_CREATE_FILE,
+                       fmt::format("Failed to open file {}", filename)));
   }
 
-  // write metadata
-  file.writeCompound("Metadata");
-  {
-    file.writeInt("WEOffsetX", info.WE_offset[0]);
-    file.writeInt("WEOffsetY", info.WE_offset[1]);
-    file.writeInt("WEOffsetZ", info.WE_offset[2]);
-    file.writeString("Name", info.schem_name_utf8.data());
-    file.writeString("Author", info.author_utf8.data());
-    file.writeLong("Date", info.date);
-
-    file.writeListHead("RequiredMods", NBT::String,
-                       info.required_mods_utf8.size());
+  auto write_version = [&]() {  // data version
+    file.writeInt("DataVersion", (int)this->MC_data_ver);
+  };
+  auto write_palette = [&]() {
+    file.writeCompound("Palette");
     {
-      for (const auto &str : info.required_mods_utf8) {
-        file.writeString("", str.data());
+      for (int idx = 0; idx < int(block_id_list.size()); idx++) {
+        file.writeInt(block_id_list[idx].c_str(), idx);
       }
-    }
-    // finish list
-  }  // finish compound
-  file.endCompound();
-
-  file.writeCompound("Palette");
-  {
-    for (int idx = 0; idx < int(block_id_list.size()); idx++) {
-      file.writeInt(block_id_list[idx].c_str(), idx);
-    }
-  }  // finished palette
-  file.endCompound();
-
-  file.writeListHead("BlockEntities", NBT::Compound, 0);
-
-  file.writeInt("DataVersion", (int)this->MC_data_ver);
-
-  file.writeShort("Width", x_range());
-  file.writeShort("Height", y_range());
-  file.writeShort("Length", z_range());
-
-  file.writeInt("Version", 2);
-
-  file.writeInt("PaletteMax", block_id_list.size());
+    }  // finished palette
+    file.endCompound();
+  };
+  auto write_offset = [&]() {
+    file.writeIntArrayHead("Offset", 3);
+    {
+      file.writeInt("", info.offset[0]);
+      file.writeInt("", info.offset[1]);
+      file.writeInt("", info.offset[2]);
+    }  // end array
+  };
+  auto write_shape = [&]() {
+    file.writeShort("Width", x_range());
+    file.writeShort("Height", y_range());
+    file.writeShort("Length", z_range());
+  };
 
   std::vector<uint8_t> blockdata;
   ::shrink_bytes_weSchem(xzy.data(), xzy.size(), block_id_list.size(),
                          &blockdata);
+  auto write_blocks = [&](const char *key) {
+    file.writeByteArrayHead(key, blockdata.size());
+    {
+      const int8_t *data = reinterpret_cast<int8_t *>(blockdata.data());
+      for (int64_t idx = 0; idx < int64_t(blockdata.size()); idx++) {
+        file.writeByte("", data[idx]);
+      }
+    }  // end array
+  };
 
-  file.writeByteArrayHead("BlockData", blockdata.size());
-  {
-    const int8_t *data = reinterpret_cast<int8_t *>(blockdata.data());
-    for (int64_t idx = 0; idx < int64_t(blockdata.size()); idx++) {
-      file.writeByte("", data[idx]);
+  if (this->MC_major_ver <= SCL_gameVersion::MC19) {
+    // write metadata
+    file.writeCompound("Metadata");
+    {
+      file.writeInt("WEOffsetX", info.WE_offset[0]);
+      file.writeInt("WEOffsetY", info.WE_offset[1]);
+      file.writeInt("WEOffsetZ", info.WE_offset[2]);
+      file.writeString("Name", info.schem_name_utf8.data());
+      file.writeString("Author", info.author_utf8.data());
+      file.writeLong("Date", info.date);
+
+      file.writeListHead("RequiredMods", NBT::String,
+                         info.required_mods_utf8.size());
+      {
+        for (const auto &str : info.required_mods_utf8) {
+          file.writeString("", str.data());
+        }
+      }
+      // finish list
+    }  // finish compound
+    file.endCompound();
+
+    file.writeInt("Version", 2);  // schematic format version
+    write_version();
+
+    write_palette();
+    file.writeInt("PaletteMax", block_id_list.size());
+
+    file.writeListHead("BlockEntities", NBT::Compound, 0);
+
+    write_shape();
+    write_offset();
+    write_blocks("BlockData");
+  } else {  // 1.20+
+    file.writeCompound("Schematic");
+    {
+      file.writeCompound("Metadata");
+      {
+        file.writeLong("Date", info.date);
+
+        file.writeCompound("WorldEdit");
+        {
+          file.writeString("Version", "unknown");
+          file.writeString("EditingPlatform", "enginehub:fabric");
+          file.writeIntArrayHead("Origin", 3);
+          {
+            for (int i = 0; i < 3; i++) {
+              file.writeInt("", 0);
+            }
+          }
+        }
+        file.endCompound();
+      }
+      file.endCompound();  // finish metadata
+
+      file.writeInt("Version", 3);  // schematic format version
+      write_version();
+      write_shape();
+      write_offset();
+
+      file.writeCompound("Blocks");
+      {
+        write_palette();
+        write_blocks("Data");
+        file.writeListHead("BlockEntities", NBT::tagType::Compound, 0);
+      }
+      file.endCompound();  // finish Blocks
     }
-  }  // end array
-
-  file.writeIntArrayHead("Offset", 3);
-  {
-    file.writeInt("x", info.offset[0]);
-    file.writeInt("y", info.offset[1]);
-    file.writeInt("z", info.offset[2]);
-  }  // end array
+    file.endCompound();
+  }
 
   file.close();
-  if (error_flag != nullptr) {
-    *error_flag = SCL_errorFlag::NO_ERROR_OCCUR;
-  }
-  if (error_str != nullptr) {
-    *error_str = "";
-  }
-  return true;
+  return {};
 }
